Use simplified Workout language
After exploring alternatives, I realized that the idea of writting
down the workouts in a structured manner (ie: YAML) would be quite
hard to sell to an average user. Mostly because we already have
been doing this without any problem on a simplistic fashion, like
a chalk board or a reddit post where you describe an amount (reps,
dintance or time) then exercise to be performed.

Based on that, I am taking the long road by creating a parser for
a simple "workout" language. I agree a formal language might be a
little rigid for this context but it worth to try (plus my AI
knowledge on natural language processing is spare).