Add support for workout metadata
I could say the parser is already done by recognizing titles, cir-
cuits, sets, amounts and exercises; BUT there is plenty meta data
the user migh require to add, like extra rules or limits.

From past experiences, I recall on the gym's workouts boards there
was a sense of metadata, which most of the time was wrote on the
title (or close to it). Based on that observation I chose to
extract the metadata that could within any workout/circuit title.