Refactor metadata recognition
The first approach for title metadata recognition was not working
as expected, so I had to switch a couple of things around.

I dropped the // for comments and stick with what Markdown does
(hint: HTML comment). Then tune a little bit the stateful lexer
for greedy title lines so the grammar could be a little bit
simpler.